Moved javaScriptAlertPage.js to the correct place
