+++ cypress/Integration/PageObject/javaScriptAlertPage.js
